+++ src/lib/app-provider/hooks/index.ts
@@ -7,3 +7,5 @@ export * from "./useExecuteCmds";
 export * from "./useTokensInfo";
 export * from "./useInternalNavigate";
 export * from "./useNetworkChange";
+export * from "./useSelectChain";
+export * from "./useCurrentNetwork";