+++ src/lib/pages/deploy/index.tsx
@@ -1,12 +1,28 @@
-import { Heading, Text } from "@chakra-ui/react";
+import {
+  Alert,
+  AlertDescription,
+  AlertIcon,
+  Flex,
+  Heading,
+  Text,
+} from "@chakra-ui/react";
 
-import { useInternalNavigate } from "lib/app-provider";
+import {
+  useCurrentNetwork,
+  useInternalNavigate,
+  useSelectChain,
+} from "lib/app-provider";
 import { ButtonCard } from "lib/components/ButtonCard";
 import { Stepper } from "lib/components/stepper";
 import WasmPageContainer from "lib/components/WasmPageContainer";
+import { getChainNameByNetwork } from "lib/data";
 
 const Deploy = () => {
+  const network = useCurrentNetwork();
   const navigate = useInternalNavigate();
+  const selectChain = useSelectChain();
+
+  const isMainnet = network === "mainnet";
   return (
     <WasmPageContainer>
       <Text variant="body1" color="text.dark" mb={3} fontWeight={700}>
@@ -16,9 +32,38 @@ const Deploy = () => {
       <Heading as="h4" variant="h4" my="48px">
         Select Deploy Option
       </Heading>
+      {isMainnet && (
+        <Alert variant="violet" mb="16px" alignItems="flex-start">
+          <AlertIcon mt={2} />
+          <AlertDescription>
+            Uploading new Wasm files on permissioned chains is coming soon to
+            Celatone. Currently, you can upload codes and instantiate contracts
+            without permission on testnet.
+          </AlertDescription>
+        </Alert>
+      )}
       <ButtonCard
         title="Upload new WASM File"
-        description="Store a new Wasm file on-chain"
+        description={
+          isMainnet ? (
+            <Flex fontSize="14px" gap={1}>
+              <Text color="text.disabled">
+                Currently available on testnet only.
+              </Text>
+              <Text
+                color="honeydew.main"
+                _hover={{ textDecoration: "underline" }}
+                cursor="pointer"
+                onClick={() => selectChain(getChainNameByNetwork("testnet"))}
+              >
+                Switch to testnet
+              </Text>
+            </Flex>
+          ) : (
+            "Store a new Wasm file on-chain"
+          )
+        }
+        disabled={isMainnet}
         onClick={() => navigate({ pathname: "/upload" })}
         mb="16px"
       />